Changed sorting algorithm from bottom up approach
Rereading the problem, it appeared that the solution I presented didn't match the original criterion. Solution called for an approach that first point is the top left point of the grid and rows read right to left, top to bottom as though you were reading the points off the grid. Columns are read as the rows reading down. This can be done by instead of flipping the grid, simply re-arranging the points.